add fees
