+++ src/hooks/index.ts
@@ -6,6 +6,7 @@ export * from "./useCompounderHelpers"
 export * from "./useCompounderPreviewDeposit"
 export * from "./useCompounderPreviewRedeem"
 export * from "./useCompounderVault"
+export * from "./useCompounderVaultFees"
 export * from "./useConcentratorApy"
 export * from "./useConcentratorAum"
 export * from "./useConcentratorHelpers"
